# Improve IsFQDN performance