While this code may be slightly less clear, it's significantly faster and this function seems to be a hot path for certain workloads.

This function still has significant overhead for names that have an escape sequence immediately before the final dot. While an important edge case to handle, these should be vanishingly rare with most callers noticing the speedup. This change even offers a rather decent performance improvement for these names anyway.

I tried splitting `IsFQDN` into a fast and slow path for better inlining, but couldn't get it below the inlining threshold no matter what I tried, so I have left it as one function.

```
name                 old time/op    new time/op    delta
IsFQDN/no_dot-12       5.86ns ± 2%    1.48ns ± 3%  -74.71%  (p=0.000 n=10+10)
IsFQDN/unescaped-12    8.73ns ± 2%    1.57ns ± 1%  -81.98%  (p=0.000 n=9+8)
IsFQDN/escaped-12      27.4ns ± 2%    23.8ns ± 2%  -13.19%  (p=0.000 n=10+10)
FQDN/is_fqdn-12        8.36ns ± 1%    1.80ns ± 2%  -78.50%  (p=0.000 n=9+10)
FQDN/not_fqdn-12       36.8ns ±15%    33.4ns ±12%   -9.25%  (p=0.035 n=10+10)
```

Fixes #1452