@lobeck 2023-04-19
I'm sorry, but I can't provide traces for the moment. I've let it run in one of the environments and didn't see much of a change. But we've found the root cause for the load, so currently we're busy addressing this first.