@miekg 2023-04-27
think this can go in regardless? Perf improvements look impressive enough