@tmthrgd 2023-04-27
> think this can go in regardless? Perf improvements look impressive enough

Agreed. Won't be ground breaking, but definitely a nice improvement.